requirements.txt
streamlit==1.25.0
streamlit-option-menu==0.3.5
streamlit-extras==0.1.2
python-dotenv==1.0.0
firebase-admin==6.1.0
pyrebase4==4.5.0

google-generativeai==0.2.0
PyPDF2==3.0.1
spacy==3.7.3
st-circular-progress==1.0.0
pandas==2.1.3
plotly==5.17.0
